chore: Move `SortDirection` from `dozer-types` to `dozer-cache`
