+++ dozer-cache/src/cache/expression/tests/serialize.rs
@@ -1,11 +1,11 @@
 use crate::cache::expression::FilterExpression;
 use crate::cache::expression::Operator;
+use crate::cache::expression::SortDirection::{Ascending, Descending};
 use crate::cache::expression::SortOption;
 use crate::cache::expression::SortOptions;
 use dozer_types::serde_json;
 use dozer_types::serde_json::json;
 use dozer_types::serde_json::Value;
-use dozer_types::types::SortDirection::{Ascending, Descending};
 
 #[test]
 fn test_serialize_filter_simple() {
